Rename service to Payment Service
