+++ packages/point-of-sale/src/payments/service.ts
@@ -1,8 +1,8 @@
 import { Logger } from "pino";
-import { CREATE_INCOMING_PAYMENT } from "./mutations/createIncomingPayment";
+import { CREATE_INCOMING_PAYMENT } from "../graphql/mutations/createIncomingPayment";
 import { IAppConfig } from "../config/app";
 import { ApolloClient, NormalizedCacheObject } from '@apollo/client'
-import { AmountInput, CreateIncomingPaymentInput, type Mutation } from "./generated/graphql";
+import { AmountInput, CreateIncomingPaymentInput, type Mutation } from "../graphql/generated/graphql";
 import { FnWithDeps } from "../shared/types";
 import { v4 } from "uuid";
 
@@ -12,13 +12,13 @@ type ServiceDependencies = {
   apolloClient: ApolloClient<NormalizedCacheObject>
 }
 
-type GraphQLService = {
+type PaymentService = {
   createIncomingPayment: (walletAddressId: string, incomingAmount: AmountInput) => Promise<string>
 }
 
-export function createGraphQLService(deps_: ServiceDependencies): GraphQLService {
+export function createPaymentService(deps_: ServiceDependencies): PaymentService {
   const logger = deps_.logger.child({
-    service: 'GraphQLServce'
+    service: 'PaymentService'
   })
   const deps = {
     ...deps_,
@@ -31,7 +31,7 @@ export function createGraphQLService(deps_: ServiceDependencies): GraphQLService
   }
 }
 
-const createIncomingPayment: FnWithDeps<ServiceDependencies, GraphQLService["createIncomingPayment"]> = async (deps, walletAddressId, incomingAmount) => {
+const createIncomingPayment: FnWithDeps<ServiceDependencies, PaymentService["createIncomingPayment"]> = async (deps, walletAddressId, incomingAmount) => {
   const client = deps.apolloClient;
   const { data } = await client.mutate<Mutation["createIncomingPayment"], CreateIncomingPaymentInput>({
     mutation: CREATE_INCOMING_PAYMENT,
